+++ tests/res/crash.rs
@@ -1,10 +1,10 @@
-#[cfg(feature = "custom-transport")]
+#[cfg(feature = "transport-custom")]
 #[path = "../util/custom_transport.rs"]
 #[rustfmt::skip]
 mod custom_transport;
 
 use anyhow::Result;
-#[cfg(feature = "custom-transport")]
+#[cfg(feature = "transport-custom")]
 use custom_transport::Transport;
 use libloading::{Library, Symbol};
 use sentry::{Breadcrumb, Options, User};
@@ -52,7 +52,7 @@ async fn main() -> Result<()> {
         "test attachment",
         "C:/rust/sentry-contrib-native/tests/res/attachment.txt",
     );*/
-    #[cfg(feature = "custom-transport")]
+    #[cfg(feature = "transport-custom")]
     options.set_transport(Transport::new);
     let _shutdown = options.init()?;