+++ CHANGELOG.md
@@ -17,6 +17,8 @@ and this project adheres to
 - Replaced `user_consent_give`, `user_consent_revoke` and `user_consent_reset`
   with `set_user_consent`.
 - Renamed `user_consent_get` to `user_consent`.
+- Renamed feature `default-transport` to `transport-default` and
+  `custom-transport` to `transport-custom`.
 
 ### Deprecated